Create tictactoe
Tic tac toe is developed in python. Game overview is 
1-first we initilize board
2- we display board
3- we declare module play game and call all the functions. we called
a- handle the turn
b- check win or not
c- flip the player
d- in last we check outside the loop wheather game ended or not

in game we declared global variable winner which is being called in all the functions given below.